feat: add hpx dot1d, tests, and benchmark
