+++ src/hpyx/runtime.py
@@ -23,10 +23,10 @@ class HPXRuntime:
             f"hpx.diagnostics_on_terminate!={int(diagnostics_on_terminate)}",
             f"hpx.parcel.tcp.enable!={int(tcp_enable)}",
         ]
-        
+
         if os_threads != "auto":
             cfg.append(f"hpx.os_threads!={os_threads}")
-        
+
         hpyx._core.init_hpx_runtime(cfg)
 
     def __enter__(self):
